Update the Token Team Attribute

Update the custom API Token to select the team to pull from. Must be saved as the token values ie token <- EF_UpdateTokenTeam(token, 1)

Usage

EF_UpdateTokenTeam(token, TEAM_ID)

Arguments

token

API Token Object (see EF_CreateAPIToken

TEAM_ID

Team Id referenced in EF_GetTeamsList

Value

Custom API Token Object

Examples

if (FALSE) { # \dontrun{
#Intialize Token
token <- EF_CreateAPIToken("https://your.eliteform.com", "your-api-key")

# Get Team ID
Team_list <- EF_GetTeamsList(token)

# Update Token with Team ID
token <- EF_UpdateTokenTeam(token, TEAM_ID$TeamId)
} # }